\name{print.clinsig}
\alias{print.clinsig}
\title{Display a clinsig list}
\description{
 Displays the result of the \code{clinsig} function.
}
\usage{
 \method{print}{clinsig}(x,...)
}
\arguments{
  \item{x}{a clinsig table produced by \code{\link{clinsig}}}
  \item{...}{additional arguments passed to \code{print}.}
}
\details{
 \code{print.clinsig} displays a list produced by \code{clinsig}.
}
\value{nil}
\author{Jim Lemon}
\seealso{\code{\link{clinsig}}}
\keyword{misc}
