\name{print.clinsig}
\alias{print.clinsig}
\title{Display a clinsig list}
\description{
 Displays the result of the \samp{clinsig} function.
}
\usage{
 \method{print}{clinsig}(x,...)
}
\arguments{
  \item{x}{a clinsig table produced by \samp{\link{clinsig}}}
  \item{...}{additional arguments passed to \samp{print}.}
}
\details{
 \samp{print.clinsig} displays a list produced by \samp{clinsig}.
}
\value{nil}
\author{Jim Lemon}
\seealso{\samp{\link{clinsig}}}
\keyword{misc}
